fix: enable fast render and fix tests
