+++ src/components/common/tab-panel/tab-panel-helper.ts
@@ -1,0 +1,6 @@
+const a11yProps = (index: number) => ({
+  id: `simple-tab-${index}`,
+  "aria-controls": `simple-tabpanel-${index}`,
+});
+
+export default a11yProps;
