+++ src/components/resume/work/TechBarChart.tsx
@@ -8,15 +8,9 @@ import {
   Tooltip,
 } from "recharts";
 import { useTheme } from "@mui/material/styles";
-import { Payload } from "recharts/types/component/DefaultTooltipContent";
 import dateObj from "../../../apis/DateHelper";
 import languageExp from "../../../constants/languages";
-
-type Months = string | number;
-
-interface TooltipPayload extends Payload<string, string> {
-  payload?: { name: string };
-}
+import { tooltipFormatter } from "./job-helpers";
 
 const data = languageExp
   .map((obj) => ({
@@ -24,21 +18,6 @@ const data = languageExp
     months: dateObj(obj.end).diff(obj.start, "month"),
   }))
   .sort((a, b) => b.months - a.months);
-
-export const tooltipFormatter = (
-  months: Months,
-  _name: string,
-  entry: TooltipPayload
-): [string, string] => {
-  const displayMonths = (months as number) % 12;
-  const years = Math.floor((months as number) / 12);
-
-  const label = entry.payload?.name || "";
-  const value =
-    (years ? `${years}y ` : "") + (displayMonths ? `${displayMonths}m` : "");
-
-  return [value, label];
-};
 
 const TechBarChart: React.FC = React.memo(() => {
   const {
